requirements.txt
flask
flask_sqlalchemy
sqlalchemy.exc
flask-restx
psycopg2-binary
gunicorn